Refactor codes using plugindata
