+++ src/coolqbot/config.py
@@ -11,14 +11,10 @@ HOME_DIR_PATH = Path(__file__).parents[1]
 if platform.system() == 'Linux':
     CONFIG_PATH = Path('/home/user/coolq/bot/bot.conf')
     LOG_FILE_PATH = Path('/home/user/coolq/bot/bot.log')
-    RECORDER_FILE_PATH = Path('/home/user/coolq/bot/recorder.pkl')
-    HISTORY_DIR_PATH = Path('/home/user/coolq/bot/history')
     DATA_DIR_PATH = Path('/home/user/coolq/bot/data')
 else:
     CONFIG_PATH = Path('bot.conf')
     LOG_FILE_PATH = Path('bot.log')
-    RECORDER_FILE_PATH = Path('recorder.pkl')
-    HISTORY_DIR_PATH = HOME_DIR_PATH / 'history'
     DATA_DIR_PATH = HOME_DIR_PATH / 'data'
 
 PLUGINS_DIR_PATH = HOME_DIR_PATH / 'plugins'
