+++ src/plugins/history.py
@@ -5,10 +5,24 @@ from datetime import datetime, timedelta
 from dateutil.relativedelta import relativedelta
 
 from coolqbot.bot import bot
-from coolqbot.config import HISTORY_DIR_PATH
-from coolqbot.recorder import Recorder
-from coolqbot.utils import get_history_pkl_name
+from coolqbot.plugin import PluginData
+from coolqbot.utils import get_history_pkl_name, scheduler
 from plugins.rank import get_repeat_number_ranking, get_repeat_rate_ranking
+from plugins.recorder import Recorder, recorder
+
+historydata = PluginData('history')
+
+
+@scheduler.scheduled_job('cron', day=1, hour=0, minute=0, second=0)
+async def clear_data():
+    """ 每个月最后24点(下月0点)保存记录于历史记录文件夹，并重置记录
+    """
+    # 保存数据到历史文件夹
+    date = datetime.now() - timedelta(hours=1)
+    historydata.save_pkl(recorder.get_data(), get_history_pkl_name(date))
+    # 清除现有数据
+    recorder.clear_data()
+    bot.logger.info('记录清除完成')
 
 
 @bot.on_message('group', 'private')
@@ -40,21 +54,20 @@ async def history(context):
         else:
             date = datetime.now() - relativedelta(months=1)
 
-        history_file = HISTORY_DIR_PATH / f'{get_history_pkl_name(date)}.pkl'
-        bot.logger.debug(history_file)
-
-        if not history_file.exists():
+        # 尝试读取历史数据
+        history_filename = f'{get_history_pkl_name(date)}.pkl'
+        if not historydata.exists(history_filename):
             str_data = f'{date.year}年{date.month}月数据不存在，请换试试吧0.0'
             return {'reply': str_data, 'at_sender': False}
 
         # 如无其他情况，开始获取历史记录，并输出排行榜
-        history_data = Recorder(history_file)
+        history_data = historydata.load_pkl(history_filename)
         display_number = 10000
         minimal_msg_number = 0
         display_total_number = True
 
-        repeat_list = history_data.repeat_list
-        msg_number_list = history_data.msg_number_list
+        repeat_list = history_data['repeat_list']
+        msg_number_list = history_data['msg_number_list']
 
         repeat_rate_ranking = await get_repeat_rate_ranking(repeat_list, msg_number_list, display_number, minimal_msg_number, display_total_number)
         repeat_number_ranking = await get_repeat_number_ranking(repeat_list, display_number)
