+++ src/plugins/rank.py
@@ -5,7 +5,7 @@ from operator import itemgetter
 
 from coolqbot.bot import bot
 from coolqbot.config import GROUP_ID
-from coolqbot.recorder import recorder
+from plugins.recorder import recorder
 
 
 @bot.on_message('group', 'private')
@@ -32,7 +32,7 @@ async def rank(context):
 
         str_data = ''
         repeat_rate_ranking = await get_repeat_rate_ranking(repeat_list, msg_number_list, display_number, minimal_msg_number, display_total_number)
-        repeat_number_ranking = await get_repeat_number_ranking(repeat_list, display_number)
+        repeat_number_ranking = await get_repeat_number_ranking(repeat_list, msg_number_list, display_number, display_total_number)
 
         if repeat_rate_ranking and repeat_rate_ranking:
             str_data = repeat_rate_ranking + '\n\n' + repeat_number_ranking
@@ -42,7 +42,7 @@ async def rank(context):
         return {'reply': str_data, 'at_sender': False}
 
 
-async def get_repeat_number_ranking(repeat_number_list, x):
+async def get_repeat_number_ranking(repeat_number_list, msg_number_list, x, display_total_number):
     '''获取次数排行榜'''
     od = collections.OrderedDict(
         sorted(repeat_number_list.items(), key=itemgetter(1), reverse=True))
@@ -50,7 +50,10 @@ async def get_repeat_number_ranking(repeat_number_list, x):
     str_data = '复读次数排行榜'
     for k, v in od.items():
         name = await get_name(k)
-        str_data += f'\n{name}: {v}次'
+        if display_total_number:
+            str_data += f'\n{name}({msg_number_list[k]}): {v}次'
+        else:
+            str_data += f'\n{name}: {v}次'
         i += 1
         if i > x:
             break
